Run prettier

+++ contracts/src/v0.8/interfaces/ChainlinkPoRAddressList.sol
@@ -1,6 +1,7 @@
 pragma solidity ^0.8.0;
 
 interface ChainlinkPoRAddressList {
-    function getPoRAddressListLength() external view returns (uint);
-    function getPoRAddressList(uint startIndex, uint endIndex) external view returns (string[] memory);
+  function getPoRAddressListLength() external view returns (uint256);
+
+  function getPoRAddressList(uint256 startIndex, uint256 endIndex) external view returns (string[] memory);
 }
